requirements.txt
python-telegram-bot[standard]==20.8
python-dotenv